fix(ci): clear the two clippy errors blocking Rust CI
Rust CI has failed on main for the last four runs, so every PR inherits a red
`build` check. Both errors are `-D warnings` promotions from a newer clippy and
are unrelated to any recent change:

  api-gateway/src/handlers/websocket.rs:92
    useless conversion to the same type: `String`  (drop `.into()`)

  analysis-engine/src/analyzers/static_analyzer.rs:707
    casting to the same type is unnecessary (`u32` -> `u32`)  (drop the cast)

Fixed here rather than in a separate PR because they block this one from
merging; neither file is otherwise touched by this branch.

+++ backend/analysis-engine/src/analyzers/static_analyzer.rs
@@ -704,7 +704,7 @@ impl StaticAnalyzer {
         Ok(PEAnalysis {
             machine_type: format!("{:?}", pe.header.coff_header.machine),
             timestamp: Some(pe.header.coff_header.time_date_stamp),
-            entry_point: pe.entry as u32,
+            entry_point: pe.entry,
             sections,
             imports,
             exports,
